add: board, todoList 완성

+++ src/main/java/com/nbacm/trelldochi/domain/list/dto/TodoListRequestDto.java
@@ -9,11 +9,11 @@ import java.util.List;
 public class TodoListRequestDto {
 
     private String title;
-    private Long listOrder;
+    private int listOrder;
     private List<Card> cardList;
     private Long boardId;
 
-    public TodoListRequestDto(String title, Long listOrder, List<Card> cardList, Long boardId) {
+    public TodoListRequestDto(String title, int listOrder, List<Card> cardList, Long boardId) {
         this.title = title;
         this.listOrder = listOrder;
         this.cardList = cardList;
